+++ et-driver/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - [CI] Adapt to conan pre-release versioning
 - conan: remove bundle computation to resolve the conflict
+- [SW-18673] Allow reinitiate of device when device appears to be initialized. The use-case is when device state is changed from the backdoor such as PMIC UART, this can help reinitiate the device
 ### Deprecated
 ### Removed
 ### Fixed